Merged changes from 3.6.5.8
Original commit: 3.6.5.8

At the end of the day almost all of the issues with the code not building on Mac are resolved, the only remaining differences are Xcode project changes, that one thing in an SDL include, and the undoing of the new way to figure out OpenGL version. Everything else is fine.

Mac still builds and runs